+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * added `stubbles\peer\IpAddress::isValid()`
   * added `stubbles\peer\http\HttpUri::isValid()`
   * added `stubbles\peer\http\HttpUri::exists()`
+  * added integration with `stubbles\values\Value` when present
 
 
 7.0.0 (2016-01-11)